change in algorithm and structure; increased delays; 24 hours pause upon reaching 300 visits
old algo:
collect all urls from all companies on the file from level 1 to level 5
or 6 then remove duplicates then visit the unique urls, get info then
save

new algo:
getting info on the first page visit then compare if the info is already
on the file. If yes ignore if no save to file.

increased the delays from 5 to 30 seconds to 15 to 45 seconds; pause for
24 hours after visiting 300 company page